Got rid of CacheNotificationVerticle
