+++ src/main/java/org/prebid/server/spring/config/VertxConfiguration.java
@@ -3,6 +3,7 @@ package org.prebid.server.spring.config;
 import io.vertx.core.Vertx;
 import io.vertx.core.VertxOptions;
 import io.vertx.core.file.FileSystem;
+import io.vertx.ext.web.handler.BodyHandler;
 import org.prebid.server.vertx.ContextRunner;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -19,6 +20,11 @@ public class VertxConfiguration {
     @Bean
     FileSystem fileSystem(Vertx vertx) {
         return vertx.fileSystem();
+    }
+
+    @Bean
+    BodyHandler bodyHandler(@Value("${vertx.uploads-dir}") String uploadsDir) {
+        return BodyHandler.create(uploadsDir);
     }
 
     @Bean
